Move notification clases
